libgo: 1.25 reflect.TypeFor + net sendFile reconciliation
reflect.TypeFor: restore gccgo idiom (var v T; TypeOf(v); *T for interfaces)
instead of gc's toType(abi.TypeFor[T]()) — gccgo reflect uses its own *rtype,
not *abi.Type, so the gc form fails to type-check when TypeFor[T] is
instantiated (broke json/xml/asn1/database-sql/text-template/x-tools).

net: exclude gc's new sendfile_nonwindows.go on gccgo (&& !gccgo); it
redefined sendFile, which gccgo provides via sendfile_glibc.go. Fix
sendfile_glibc.go to pass uintptr(fd) to poll.SendFile (signature is now
uintptr).

+++ libgo/go/net/sendfile_glibc.go
@@ -41,7 +41,7 @@ func sendFile(c *netFD, r io.Reader) (written int64, err error, handled bool) {
 
 	var werr error
 	err = sc.Read(func(fd uintptr) bool {
-		written, werr, handled = poll.SendFile(&c.pfd, int(fd), remain)
+		written, werr, handled = poll.SendFile(&c.pfd, uintptr(fd), remain)
 		return true
 	})
 	if err == nil {
